Move `RangeCollector` and restore substring module
