+++ tlsn/tlsn-prover/src/tls/state.rs
@@ -14,14 +14,12 @@ use mpz_garble::{
 use mpz_share_conversion::{ConverterSender, Gf2_128};
 use tls_core::{handshake::HandshakeData, key::PublicKey};
 use tls_mpc::MpcTlsLeader;
-use tlsn_core::{
-    commitment::TranscriptCommitmentBuilder, proof::DirectSubstringsProofBuilder, Transcript,
-};
+use tlsn_core::{commitment::TranscriptCommitmentBuilder, Transcript};
 use utils::range::RangeSet;
 
 use crate::{
     tls::{MuxFuture, OTFuture},
-    Mux,
+    Mux, RangeCollector,
 };
 
 use super::ProverError;
@@ -131,7 +129,7 @@ pub struct Verify {
     pub(crate) server_public_key: PublicKey,
     pub(crate) transcript_tx: Transcript,
     pub(crate) transcript_rx: Transcript,
-    pub(crate) proof_builder: DirectSubstringsProofBuilder,
+    pub(crate) proof_builder: RangeCollector,
 }
 
 impl From<Closed> for Verify {
@@ -147,7 +145,7 @@ impl From<Closed> for Verify {
             server_public_key: state.server_public_key,
             transcript_tx: state.transcript_tx,
             transcript_rx: state.transcript_rx,
-            proof_builder: DirectSubstringsProofBuilder::default(),
+            proof_builder: RangeCollector::default(),
         }
     }
 }
